[REF] website_event(_track): improve tests and test classes
RATIONALE

Events are sometimes held online, gathering a community. In this merge
we improve Event application to better support full-online events with
improved tracks, wishlists, chat rooms, ...

SPECIFICATIONS

Improve website_event and website_event_track test coding style, notably

  * use specific users when performing tests to trigger access rights issues
    (this notably allowed to find an issue with event groups not allowed to
    update menus);
  * define common classes;

LINKS

Task ID-2252655 (Main Online Event task)
Task ID-2283796 (Event B2Basics)
