+++ addons/website_event/tests/__init__.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
+from . import common
+from . import test_event_internals
 from . import test_ui
-from . import test_event_website
+from . import test_website_event
