+++ addons/website_event_track/tests/__init__.py
@@ -1,1 +1,4 @@
-from . import test_event_track_website
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+from . import test_website_event
